# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/pengunjung": {
            "get": {
                "tags": [
                    "Pengunjung Perpustakaan"
                ],
                "summary": "Get pengunjung",
                "operationId": "51257e27427a4a399660808722ea6a8d",
                "responses": {
                    "200": {
                        "description": "Get pengunjung"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengunjung Perpustakaan"
                ],
                "summary": "Post pengunjung",
                "operationId": "bdecbef775c3dbb9aa8463c6ca75a1ba",
                "responses": {
                    "200": {
                        "description": "Post pengunjung"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/qrcode": {
            "get": {
                "tags": [
                    "Pengunjung Perpustakaan"
                ],
                "summary": "Get qrcode",
                "operationId": "5ac4f4edab8a03963750d0a210da0e64",
                "responses": {
                    "200": {
                        "description": "Get qrcode"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/jumlah-dokumen": {
            "get": {
                "tags": [
                    "Statistik"
                ],
                "summary": "Get jumlah-dokumen",
                "operationId": "b0e862743ed77a63bf45b54ef6f2b7b1",
                "responses": {
                    "200": {
                        "description": "Get jumlah-dokumen"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/jumlah-peminjaman-dokumen": {
            "get": {
                "tags": [
                    "Statistik"
                ],
                "summary": "Get jumlah-peminjaman-dokumen",
                "operationId": "cbb9b975140e4dace365991dec7a6df4",
                "responses": {
                    "200": {
                        "description": "Get jumlah-peminjaman-dokumen"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/jumlah-ruangan": {
            "get": {
                "tags": [
                    "Statistik"
                ],
                "summary": "Get jumlah-ruangan",
                "operationId": "a354908d1909b96cc144b42f6f1c8e5b",
                "responses": {
                    "200": {
                        "description": "Get jumlah-ruangan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/jumlah-peminjaman-ruangan": {
            "get": {
                "tags": [
                    "Statistik"
                ],
                "summary": "Get jumlah-peminjaman-ruangan",
                "operationId": "3ae675ae137e49b975947a8514f672c3",
                "responses": {
                    "200": {
                        "description": "Get jumlah-peminjaman-ruangan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/jumlah-pengunjung": {
            "get": {
                "tags": [
                    "Statistik"
                ],
                "summary": "Get jumlah-pengunjung",
                "operationId": "f2e0b7ecd9bdb4d88e9f7a63ee7cc01f",
                "responses": {
                    "200": {
                        "description": "Get jumlah-pengunjung"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}